+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - fix(ipv6): unspecified IPv6 target hostname (::)"
 - fix(response-interceptor): reduce responseInterceptor buffer churn
 - fix(ws): handle multi-server upgrade subscription and safe proxy shutdown
+- feat(router): add 'res' and 'options' to router function
 
 ## [v4.0.0](https://github.com/chimurai/http-proxy-middleware/releases/tag/v4.0.0)